% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar_matriz_conf.R
\name{graficar_matriz_conf}
\alias{graficar_matriz_conf}
\title{Graficar Matriz de Confusión con ggplot2}
\usage{
graficar_matriz_conf(conf_matrix, titulo)
}
\arguments{
\item{conf_matrix}{Objeto de clase \code{confusionMatrix} generado con \code{caret}.}

\item{titulo}{Cadena de texto con el título del gráfico.}
}
\value{
Un objeto \code{ggplot} que contiene el gráfico de la matriz de confusión.
}
\description{
Esta función genera un gráfico a partir de una matriz de confusión producida con el paquete \code{caret},
diferenciando visualmente los aciertos (en verde) de los errores (en rojo). Se normaliza la intensidad de los colores
para resaltar la magnitud de las frecuencias.
}
\examples{
\dontrun{
library(caret)
# Ejemplo de uso:
predictions <- factor(c("A", "B", "A", "A", "B"))
references  <- factor(c("A", "A", "A", "B", "B"))
conf_matrix <- confusionMatrix(predictions, references)
graficar_matriz_conf(conf_matrix, "Matriz de Confusión")
}

}
